# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trophy management API",
        "description": "API to manage trophies",
        "contact": {
            "email": "julien.raedler@twofold.swiss"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Trophy Manager API Server"
        }
    ],
    "paths": {
...
        "/trophies": {
            "get": {
                "tags": [
                    "Trophies"
                ],
                "summary": "Get list of trophies",
                "description": "Returns list of trophies",
                "operationId": "3a3bd2932ed0973968dc34b753bb194d",
                "responses": {
                    "200": {
                        "description": "Successful operation."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trophies"
                ],
                "summary": "Store new trophies",
                "description": "Returns trophies data",
                "operationId": "161f6c0e2e31e2368d78d40f1aa4ffb7",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "type_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "ranking": {
                                        "type": "integer"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "place": {
                                        "type": "string"
                                    },
                                    "oponent": {
                                        "type": "string"
                                    },
                                    "score": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    },
                                    "club_name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "type_id": "1",
                                    "title": "Tournament",
                                    "ranking": "1",
                                    "date": "2000-01-01",
                                    "category_id": "1",
                                    "place": "New York",
                                    "oponent": "Bad Guys",
                                    "score": "80/0",
                                    "price": "50000",
                                    "club_name": "NY Sports Club",
                                    "image": "image.jpg"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "A result object.",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/trophies/sortByDate": {
            "get": {
                "tags": [
                    "Trophies"
                ],
                "summary": "Get trophies information",
                "description": "Returns trophies data",
                "operationId": "97fa2657c7772d3a51f33d15c0f71314",
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "A trophies.",
                                        "value": {
                                            "id": 1,
                                            "type_id": "1",
                                            "title": "Tournament",
                                            "ranking": "1",
                                            "date": "2000-01-01",
                                            "category_id": "1",
                                            "place": "New York",
                                            "oponent": "Bad Guys",
                                            "score": "80/0",
                                            "price": "50000",
                                            "club_name": "NY Sports Club",
                                            "image": "image.jpg",
                                            "created_at": "2022-07-29 07:53:36",
                                            "updated_at": "2022-07-29 07:53:36"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/trophies/sortByRank": {
            "get": {
                "tags": [
                    "Trophies"
                ],
                "summary": "Get trophies information",
                "description": "Returns trophies data",
                "operationId": "4387f68706465452ee294b99ece4226f",
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "A trophies.",
                                        "value": {
                                            "id": 1,
                                            "type_id": "1",
                                            "title": "Tournament",
                                            "ranking": "1",
                                            "date": "2000-01-01",
                                            "category_id": "1",
                                            "place": "New York",
                                            "oponent": "Bad Guys",
                                            "score": "80/0",
                                            "price": "50000",
                                            "club_name": "NY Sports Club",
                                            "image": "image.jpg",
                                            "created_at": "2022-07-29 07:53:36",
                                            "updated_at": "2022-07-29 07:53:36"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}